Use double for monetary/metrics; add VM uptime
Replace decimal-typed fields with double across models, services, and interfaces (Payout.Amount, VirtualMachine Cost/metrics, service method signatures and implementations). Add UptimeString and a computed TimeSpan Uptime property on VirtualMachine that parses the stored string. Update sample data in Revenue.razor to use double literals (remove 'm' suffix). These changes align numeric types for Firestore serialization and add string-backed uptime handling.

+++ providerunicore/Services/IPayoutService.cs
@@ -7,6 +7,6 @@ public interface IPayoutService
     Task<Payout?> GetByIdAsync(string id);
     Task<IEnumerable<Payout>> GetAllPayoutsAsync();
     Task<IEnumerable<Payout>> GetPayoutsByStatusAsync(string status);
-    Task<Payout> CreatePayoutAsync(decimal amount, string method);
+    Task<Payout> CreatePayoutAsync(double amount, string method);
     Task<Payout> UpdatePayoutStatusAsync(string id, string status);
 }
